requirements.txt
streamlit==1.29.0
groq==0.4.1
python-dotenv==1.0.0
requests==2.31.0
